requirements.txt
archspec==0.2.5
astc-encoder-py==0.1.12
attrs==25.4.0
brotli==1.2.0
etcpak==0.9.15
fmod_toolkit==0.1.3
fsspec==2025.12.0
lz4==4.4.5
pillow==12.1.0
pyfmodex==0.7.2
texture2ddecoder==1.0.5
UnityPy==1.24.2
